Přechod na BigDecimal

+++ src/bakalarka_0/pkg2/Point3d.java
@@ -1,14 +1,16 @@
 
 package bakalarka_0.pkg2;
 
+import java.math.BigDecimal;
+
 public class Point3d {
-    double x;
-    double y;
-    double z;
-    double t;
+    BigDecimal x;
+    BigDecimal y;
+    BigDecimal z;
+    BigDecimal t;
     int id;
     //vlastní třída pro 3D bod - drží si hlavní parametry (x,y,z), čas a ID 
-    public Point3d(double x, double y, double z,double t,int id){
+    public Point3d(BigDecimal x, BigDecimal y, BigDecimal z,BigDecimal t,int id){
         this.x=x;
         this.y=y;
         this.z=z;
